requirements.txt
notifypy
PyQt5
Pillow
requests
